+++ probeye/inference/torch_/solver.py
@@ -1,3 +1,6 @@
+# standard library
+import warnings
+
 # third party imports
 import numpy as np
 import torch as th
@@ -155,6 +158,7 @@ class PyroSolver:
                 # (non-tensor) inputs; so there needs to be a conversion which
                 # takes place here; also, the given values must be rearranged
                 # in the dict-format required by forward_model's response method
+                forward_model.inp_structure = []
                 inp = {}
                 keys = [isens.name for isens in forward_model.input_sensors] +\
                        [*forward_model.prms_def.values()]
@@ -163,13 +167,15 @@ class PyroSolver:
                         inp[key] = value.detach().numpy()
                     else:
                         inp[key] = value
+                    forward_model.inp_structure.append(len_or_one(inp[key]))
 
                 # evaluate the forward model and its jacobian for the given
                 # input parameters; this is where the forward model is evaluated
                 # in its original setup without any tensors
                 response_dict = forward_model.response(inp)
                 jac_dict = forward_model.jacobian(inp)
-                jac_numpy = forward_model.jacobian_dict_to_array(inp, jac_dict)
+                jac_numpy = forward_model.jacobian_dict_to_array(
+                    inp, jac_dict, self.problem.n_latent_prms_dim)
 
                 # now we have the forward model's response in dict format;
                 # however, in this format it cannot be processed here, so we
@@ -253,10 +259,16 @@ class PyroSolver:
                 # an input sensor of the forward model) are not required to have
                 # their gradients evaluated, so these elements will have 'False'
                 # entries in ctx.needs_input_grad
-                return_val = tuple(
-                    None if not required else grad for required, grad
-                    in zip(ctx.needs_input_grad, th.flatten(grad_total)))
-                return return_val
+                return_val = [None] * self.problem.n_latent_prms
+                j = 0
+                for i, (dim, flag) in enumerate(zip(
+                        forward_model.inp_structure, ctx.needs_input_grad)):
+                    if flag:
+                        return_val[i] = grad_total[j: j + dim]
+                        j += dim
+                    else:
+                        j += 1
+                return tuple(return_val)
 
         return self._only_values(Autograd.apply)
 
@@ -266,7 +278,7 @@ class PyroSolver:
 
         Returns
         -------
-        list[torch.Tensor]
+        torch.Tensor
             The sampled values based on the latent parameter's priors.
         """
         # even if a list is returned by this function, we initialize a dict here
@@ -283,8 +295,11 @@ class PyroSolver:
                     # (i.e. the hyperparameters) are simply constants
                     hyperprms_dict[name] = self.problem.parameters[name].value
             # this is where the parameter's sample is generated with pyro
-            pyro_parameter_samples[ref_prm] = prior_obj.sample(hyperprms_dict)
-        return [*pyro_parameter_samples.values()]
+            sample = prior_obj.sample(hyperprms_dict)
+            if sample.dim() == 0:
+                sample = sample.reshape(1)
+            pyro_parameter_samples[ref_prm] = sample
+        return th.cat(tuple(pyro_parameter_samples.values()))
 
     def loglike(self, theta):
         """
@@ -390,11 +405,17 @@ class PyroSolver:
         self.raw_results = mcmc
 
         # translate the results to a common data structure and return it
-        var_names = self.problem.get_theta_names(tex=False)
-        var_names_tex = self.problem.get_theta_names(tex=True)
+        var_names = self.problem.get_theta_names(tex=False, components=False)
+        var_names_tex = self.problem.get_theta_names(tex=True, components=False)
         name_dict = {var_name: var_name_tex for var_name, var_name_tex
                      in zip(var_names, var_names_tex)}
-        inference_data = az.from_pyro(mcmc)
+        # the following warning-filter is intended to hide the
+        # DepreciationWarning that is currently always raised in the
+        # arviz.from_pyro method due to using np.bool instead of bool; as soon
+        # as this issue is fixed within arviz, the warning-filter can be removed
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore")
+            inference_data = az.from_pyro(mcmc, log_likelihood=False)
         inference_data.rename(name_dict, groups='posterior', inplace=True)
 
         return inference_data